Create getSearchResults
This was my answers Hacker_Rank question on an exam where getSearchResults takes in an array of string that are words and an array of string that are a bunch of random letter. The code tell you if the random letters string can be remade into any of the words in the word array. It returns a double array that show that row index is the random letters given with the col index being the word that appeared in the word array